Removed UInt8 references and used Ints instead to simplify the code. Implemented the Material class and converted structures to using materials instead of rgb.

+++ NightAndRadiance/PPMWriter.swift
@@ -21,7 +21,7 @@ struct PPMWriter {
         self.height = abs(height)
     }
     
-    func write(raster: Array<Array<UInt8>>) {
+    func write(raster: Array<Array<Int>>) {
         if raster.count != width * height {
             print("Raster must match width and height inputs.")
         }
